Broke down components, implemented Adherent getting by license number
